+++ pipeline/__init__.py
@@ -1,0 +1,1 @@
+"""서빙 파이프라인 패키지."""
