+++ pipeline/retrieval/__init__.py
@@ -1,0 +1,1 @@
+"""검색(retrieval) 파이프라인 패키지."""
