+++ pipeline/retrieval/query_expansion/__init__.py
@@ -1,0 +1,47 @@
+"""Query expansion 하위 모듈 패키지."""
+
+from pipeline.retrieval.query_expansion.query_expansion import (
+    QueryExpansionError,
+    build_repair_prompt,
+    expand_clause,
+)
+from pipeline.retrieval.query_expansion.query_expansion_prompt import (
+    SYSTEM_PROMPT,
+    build_user_prompt,
+)
+from pipeline.retrieval.query_expansion.query_expansion_schema import (
+    CaseQueryExpansion,
+    ClauseQueryExpansion,
+    CounselQueryExpansion,
+    DomainScope,
+    IssueType,
+    LawQueryExpansion,
+    ReferencedLawCandidate,
+    RetrievalTerms,
+    SearchTarget,
+    SourceRoutingHint,
+    SourceType,
+)
+from pipeline.retrieval.query_expansion.retrieval_adapter import (
+    build_retrieval_payload,
+)
+
+__all__ = [
+    "CaseQueryExpansion",
+    "ClauseQueryExpansion",
+    "CounselQueryExpansion",
+    "DomainScope",
+    "IssueType",
+    "LawQueryExpansion",
+    "QueryExpansionError",
+    "ReferencedLawCandidate",
+    "RetrievalTerms",
+    "SearchTarget",
+    "SourceRoutingHint",
+    "SourceType",
+    "SYSTEM_PROMPT",
+    "build_repair_prompt",
+    "build_retrieval_payload",
+    "build_user_prompt",
+    "expand_clause",
+]
