+++ pipeline/retrieval/query_expansion/query_expansion.py
@@ -1,0 +1,3 @@
+"""query_expansion 실행 로직 진입점."""
+
+from retrieval.query_expansion import *  # noqa: F401,F403
