+++ pipeline/retrieval/query_expansion/query_expansion_prompt.py
@@ -1,0 +1,3 @@
+"""query_expansion 프롬프트 진입점."""
+
+from retrieval.query_expansion_prompt import *  # noqa: F401,F403
