+++ pipeline/retrieval/query_expansion/query_expansion_schema.py
@@ -1,0 +1,3 @@
+"""query_expansion 스키마 진입점."""
+
+from retrieval.query_expansion_schema import *  # noqa: F401,F403
