+++ pipeline/retrieval/query_expansion/retrieval_adapter.py
@@ -1,0 +1,3 @@
+"""query_expansion retrieval adapter 진입점."""
+
+from retrieval.retrieval_adapter import *  # noqa: F401,F403
