+++ tests/pipeline/query_expansion/__init__.py
@@ -1,0 +1,1 @@
+"""Query expansion 테스트/샘플 실행 패키지."""
