+++ tests/pipeline/query_expansion/test/__init__.py
@@ -1,0 +1,1 @@
+"""Query expansion pytest 패키지."""
